% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dm3_HinfI_5000}
\alias{dm3_HinfI_5000}
\title{Genomic features for dm3 genome and HinfI restriction enzyme at 5 Kbp}
\format{A data frame with 26090 rows and 6 variables:
 \describe{
   \item{chr:}{chromosome}
   \item{map:}{mappability as computed by gem}
   \item{res:}{restriction enzyme density per 1 Kbp computed by Biostrings::matchPattern()}
   \item{cg:}{cg content as computed by bedtools}
   \item{bin:}{genomic bin with the format chromosome:start_position}
   \item{pos:}{start postion of the genomic bin}
 }}
\usage{
dm3_HinfI_5000
}
\description{
A \code{data.frame} containing the mappability, restriction
enzyme density and CG proportion of the dm3 genome and
HinfI restriction enzyme in 5 Kbp bins
}
\keyword{datasets}
