% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm10_BspHI_1000}
\alias{mm10_BspHI_1000}
\title{Genomic features for mm10 genome and BspHI restriction enzyme at 1 Kbp}
\format{A data frame with 2724632 rows and 5 variables:
 \describe{
   \item{chr:}{chromosome}
   \item{map:}{mappability as computed by gem}
   \item{res:}{restriction enzyme density per 1 Kbp computed by Biostrings::matchPattern()}
   \item{cg:}{cg content as computed by bedtools}
   \item{bin:}{genomic bin with the format chromosome:start_position}
   \item{pos:}{start postion of the genomic bin}
 }}
\usage{
mm10_BspHI_1000
}
\description{
A \code{data.frame} containing the mappability, restriction
enzyme density and CG proportion of the mm10 genome and
BspHI restriction enzyme in 1 Kbp bins
}
\keyword{datasets}
